Update proton-rawinput patchset for staging 82cff8bb and move previous version to legacy
The difference it makes is too big, and the currently upstreamed bits are leading to
some usability issues in games such as Warframe and Monster Hunter World, likely due
to incompleteness.
This rebase is using our previous "known good" patchset and was verified to fix the
aforementioned issues.

We might end up keeping this alive for the staging+fshack path until proton gets rebased.