pkgbuild: Remove hash-versioned legacy patches from the source array and symlink them in srcdir directly.
Makes the sums checking much quicker and work around a possible makepkg race condition (repro steps unclear - probably depends on the machine). Since the most important and used patches are current, skipping the check for legacy patches sounds reasonable. This will also prevent the list to really grow in the future since we can now only keep checking current, master rebased patchsets.
The reason for the non hash-versioned legacy patches to stay for now is to keep track of the commit hashes affected. A readme/list containing those might be added to the patches dir in the future for easier, centralized tracking.