Proton-tkg: Fix/workaround steamclient swap undesirable behavior on 32-bit when not reverting parts of the new path handling.
For some reason the steamclient/lsteamclient swap fails (especially in some 32-bit cases, but reported once on 64-bit) when the native library name doesn't mimic the module the app expects. As a workaround, we keep two copies of both libraries to cover such cases. This allows us to drop all the reverts before 4e8fcc4.