Hotfixer: Always enable _steamvr_support on 12d33d2+ (6.12+) for now
Our patches to strip vr support out are breaking on current proton experimental tree, to be merged with proton 6.3 stable down the line.
Let's enforce it to enabled for a while so we can think about it. In the current state and since we're disabling support in the proton script itself anyway, no change in behaviour is expected.