Externalize patches handling - first pass
This is a followup to eebc1553a4a08ae710c8f52235b2502699597e30 for the main patches, which will eventually end up handled similarly to hotfixes, as external modules. The plan is to be able to store legacy patches in a different repo at some point, to externalize the ever growing nature of the wine-tkg-git repo.