Clarify build requirements for NTSync5 in customization.cfg and proton-tkg.cfg
ntsync is part of the kernel UAPI, so the presence of the
ntsync module on the build host is not required to build
Wine with ntsync support. A complete ntsync header file at
/usr/include/linux/ntsync.h (replacing, if applicable, the
incomplete stub mainlined in kernel 6.10) is sufficient.

Note also that the proposed ntsync patchset targets kernels
>= 6.8, but builds still work if the build host is running
an earlier kernel. This is most immediately relevant for CI,
as the ubuntu-latest GitHub Actions runner has kernel 6.5.